+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Adds some more logging to some tasks.
 - More defensive code around SpellTable API calls.
+- Permissions check on delete channel.
 
 ## [v6.0.1](https://github.com/lexicalunit/spellbot/releases/tag/v6.0.1) - 2021-05-11